I added validation to the recorded audio and i fixed the UserRecordings component witch was not updating the user audio accordingly
