--- src/AuthContext.js
@@ -1,27 +1,0 @@
-import React, { createContext, useState, useEffect } from "react";
-import { getAuth } from "firebase/auth";
-
-export const AuthContext = createContext();
-
-const AuthProvider = ({ children }) => {
-  const [currentUser, setCurrentUser] = useState(null);
-  const auth = getAuth();
-
-  useEffect(() => {
-    const unsubscribe = auth.onAuthStateChanged((user) => {
-      setCurrentUser(user);
-    });
-
-    // Cleanup subscription on unmount
-    return () => unsubscribe();
-  }, [auth]);
-
-  // The value passed down the context includes `currentUser` and `setCurrentUser`
-  return (
-    <AuthContext.Provider value={{ currentUser, setCurrentUser }}>
-      {children}
-    </AuthContext.Provider>
-  );
-};
-
-export default AuthProvider;
